Infer test ResourceType field from policy
The test ResourceType field in policy unittest is only used
by the policy engine to map it to the policy under test.
This commit uses one of the policy's resource type fields for that
purpose.

The test LogType field for a rule is also not used anywhere and
this commit disregards it.

Addresses #610