Create Glue tables asynchronously
When a new log source with many log types is onboarded, creating the
Glue tables/views for these log types may take longer than we want in
the course of an API request.

Thus, we create the tables asynchronously, by sending a message to the
datacatalog updater queue, handled by the datacatalog updater lambda.

Closes #1370