Prefix rule result error message with the function that raised the error
When a rule evaluation raises an error, we include the error message in the
returned RuleResult and users that test a rule from the UI or users that
observe the rule errors from rule analysis see the error message.

However, they don't have any hint on which rule function raised the error:
rule(), title() or dedup()?

This commit prefixes the error message included in the rule_engine.RuleResult
with the function that raised the error.