Fix gzip files mis-detected as ms-fontobject sometimes
The http.DetectContentType() function we are using checks for
the content type of many files. It has a bug and sometimes
mis-detects gzip for ms-fontobject.

This commit uses the checks for gzip before calling that function.
This solves the bug, but is also a performance optmization since
most log files are gzipped.