Backend: Allow users to specify multiple s3 prefix->logtypes mappings per s3 log source (#2232)
* Allow multiple s3 prefixes per s3 source

* Cleanup prefix parameter from log-onboarding CFN template

Now that users can define multiple s3 prefixes per onboarded
bucket (log source), this commit removes the prefix filter from the
`s3:GetObject` permission for the log processing role so that users
can eventually update prefixes in an s3 source without having to
update the CFN template.

"Eventually" is the key word here though. Due to backwards compatibility,
users that already have deployed a stack that includes a prefix restriction,
need to update their stack to remove it. For now, we have to prompt all users
to update their stack.

In order to allow some users to not update their CFN stack, we have to change
the current FE<>BE flow so that the backend can check whether the installed
stack needs update (by querying AWS) and return a flag to the frontend accordingly.

* Add unit test for backwards compatibility

* Fix json fields to match graphql schema

* Get unique logtypes for s3 source

* Add validation for unique prefixes per source and prefix/bucket pairs

* Narrow down prefixes in the self registered sources for Cloudtrail, GuardDuty, ALB, s3 access logs

* For s3 sources in the log processor, use the log types of the longest prefix that matches the input object key